Review: remove -Wno-deprecated flag from compilation
Since the ProcotolVersion call was removed, the only remaining use
of deprecated functionality was kSecTrustResultConfirm. Including
this is fairly silly anyways since it was removed a long time ago
so remove that as well and let the default case in the switch catch
it. This means we can remove the no-deprecated warning override.