Review: Fix backend cipherbits code to match OpenSSL implementation
The returnvalue should be number of bits in the underlying encryption,
not what was implemented. This brings in a new common function which
translates an SSLCipherSuite to a number of bits since Secure Transport
doesn't have such an API call. The implementation is very fugly but it
will be fixed at some point when we decide how and where it should be
done.