WIP: Initial stab at online checksum enabling
A first cut at being able to enable checksums in a running cluster
which was initdb'd without checksums. Checksums are enabled with
the enable_data_checksums() function, after which all pages will
gain checksums when written.

db# select enable_data_checksums();
 enable_data_checksums
-----------------------
 t
(1 row)

A backend launcher/worker process, checksumhelper, will then start
and will enumerate and walk the cluster to backfill all pages which
require checksums.  When all pages are done, the checksum version
will be turned from in-progress to on, thereby enforcing checksum
validation as per usual.

The launcher is currently not launching any workers or enumerating
the database, it's only launching itself at the moment.